+++ solidity/contracts/hooks/libs/AbstractMessageIdAuthHook.sol
@@ -61,12 +61,13 @@ abstract contract AbstractMessageIdAuthHook is
 
     /// @inheritdoc IPostDispatchHook
     function hookType() external pure virtual returns (uint8) {
-        return uint8(IPostDispatchHook.Types.ID_AUTH_ISM);
+        return uint8(IPostDispatchHook.HookTypes.ID_AUTH_ISM);
     }
 
     // ============ Internal functions ============
 
     /// @inheritdoc AbstractPostDispatchHook
+    // solhint-disable-next-line hyperlane/no-virtual-override
     function _postDispatch(
         bytes calldata metadata,
         bytes calldata message