+++ typescript/sdk/src/token/config.ts
@@ -10,9 +10,11 @@ export enum TokenType {
   collateralFiat = 'collateralFiat',
   collateralUri = 'collateralUri',
   collateralCctp = 'collateralCctp',
+  collateralEverclear = 'collateralEverclear',
   native = 'native',
   nativeOpL2 = 'nativeOpL2',
   nativeOpL1 = 'nativeOpL1',
+  ethEverclear = 'ethEverclear',
   // backwards compatible alias to native
   nativeScaled = 'nativeScaled',
 }
@@ -26,8 +28,8 @@ const isMovableCollateralTokenTypeMap = {
   [TokenType.collateralCctp]: false,
   [TokenType.collateralFiat]: false,
   [TokenType.collateralUri]: false,
-  [TokenType.collateralVault]: true,
-  [TokenType.collateralVaultRebase]: true,
+  [TokenType.collateralVault]: false,
+  [TokenType.collateralVaultRebase]: false,
   [TokenType.native]: true,
   [TokenType.nativeOpL1]: false,
   [TokenType.nativeOpL2]: false,
@@ -35,6 +37,8 @@ const isMovableCollateralTokenTypeMap = {
   [TokenType.synthetic]: false,
   [TokenType.syntheticRebase]: false,
   [TokenType.syntheticUri]: false,
+  [TokenType.ethEverclear]: false,
+  [TokenType.collateralEverclear]: false,
 } as const;
 
 export type MovableTokenType = {
@@ -42,6 +46,10 @@ export type MovableTokenType = {
     ? K
     : never;
 }[keyof typeof isMovableCollateralTokenTypeMap];
+
+export type EverclearTokenBridgeTokenType =
+  | TokenType.ethEverclear
+  | TokenType.collateralEverclear;
 
 export function isMovableCollateralTokenType(type: TokenType): boolean {
   return !!isMovableCollateralTokenTypeMap[type];